requirements.txt
# Core Django Packages
Django==5.1
django-cors-headers==4.3.1
django-filter==23.5

# Django REST Framework + Auth
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7

# PDF + OCR Tools
Pillow==10.2.0
pdfplumber==0.10.3
pytesseract==0.3.10
reportlab==4.0.7

# Utility Packages
python-decouple==3.8

# Deployment
gunicorn==21.2.0

# Extra Common Dependencies (recommended for Django projects)
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2024.1
